Refactor file paths and name (#150)
* Refactor file paths and names, and rename packages within files.


* feat(album): add error handling for album pattern parsing

This commit adds error handling for parsing the album pattern in the deleteAlbum function. If the album pattern cannot be parsed, an error message is returned. This ensures that any errors in parsing the album pattern are properly handled.


---------
