+++ cmd/upload/TEST_DATA/Takeout1/Google Photos/Album test 6-10-23/PXL_20231006_063851485.jpg
