+++ cmd/upload/TEST_DATA/Takeout2/Google Photos/Photos from 2023/PXL_20231006_063000139.jpg
