+++ cmd/upload/TEST_DATA/Takeout2/Google Photos/Sans titre(9)/PXL_20231006_063108407.jpg.json
