+++ cmd/upload/TEST_DATA/Takeout2/Google Photos/Sans titre(9)/métadonnées.json
