+++ cmd/upload/TEST_DATA/Takeout3/Photos from 2023/DSC_0238_1.JPG.json
