+++ cmd/upload/TEST_DATA/folder/high/AlbumA/PXL_20231006_063000139.jpg
