+++ cmd/upload/TEST_DATA/folder/high/AlbumB/PXL_20231006_063528961.jpg
