+++ cmd/upload/TEST_DATA/folder/high/AlbumB/PXL_20231006_063851485.jpg
