+++ cmd/upload/TEST_DATA/folder/low/PXL_20231006_063000139.jpg
