+++ cmd/upload/TEST_DATA/folder/low/PXL_20231006_063121958.jpg
