+++ cmd/upload/assets.go
@@ -1,4 +1,4 @@
-package cmdupload
+package upload
 
 import (
 	"fmt"
