+++ cmd/upload/upload_test.go
@@ -1,4 +1,4 @@
-package cmdupload
+package upload
 
 import (
 	"cmp"
